@ksss 2017-05-08
You can also see https://bugs.ruby-lang.org/issues/12373