@joshpencheon 2017-05-08
Thanks @ksss. @jeg2, is this something I can help with? I think this patch goes a little further than the linked ruby-lang issue, which hasn't seen any progress for a year.