@hsbt 2017-05-16
@joshpencheon Thank you for your patch. I'm ok to merge your pull request.

But ksss's patch was the first time, I will review your patch after applying it first.