@joshpencheon 2017-05-16
Of course @hsbt - I think that the commit https://github.com/ruby/ruby/pull/1607/commits/e8d67b6bde in this PR will complement the work of @ksss.