@ksss 2017-05-17
@joneslee85 @hsbt Thank you for advancing about this issue.
I replied to https://bugs.ruby-lang.org/issues/12373