@hsbt 2017-05-17
@joshpencheon Can you rebase and re-consider this pull request from https://github.com/ruby/ruby/commit/15cb9817a234d8b67ae412c8eabda91d46a8ef2c ?

Thank you.