@joshpencheon 2017-05-17
Thanks @hsbt , I've rebased and updated the PR. This will avoid the allocation of temporary arrays, and a new `#join`'d string per cell.