@hsbt 2017-05-18
@joshpencheon I confirmed to benchmark results with your patch. It seems ok to me.

#### benchmark-ips results

```
trunk:
Warming up --------------------------------------
                         4.000  i/100ms
Calculating -------------------------------------
                         39.661  (±10.1%) i/s -      2.352k in  60.034781s
with-patch:
Warming up --------------------------------------
                         5.000  i/100ms
Calculating -------------------------------------
                         60.521  (± 9.9%) i/s -      3.595k in  60.047157s
```

#### memory_profiler resuts

```
trunk:
allocated memory by class
-----------------------------------
  35588490  String
   7454320  Array
    294000  MatchData
     37340  Regexp
     11840  Hash
      2400  CSV
      1600  Proc
      1280  Method
       800  StringIO
with-patch:
allocated memory by class
-----------------------------------
  18788490  String
   3454320  Array
    294000  MatchData
     37340  Regexp
     11840  Hash
      2400  CSV
      1600  Proc
      1280  Method
       800  StringIO
```